Rename TasksModel; add TaskModel for adapter
